Time: 2 ms (34.8%), Space: 11.3 MB (20.67%) - LeetHub

+++ 0338-counting-bits/0338-counting-bits.cpp
@@ -6,7 +6,7 @@ public:
         {
             int cnt = 0;
             int copyOfI = i;
-            for (short x = 0; i>0 && x < 32 ; x++)
+            for (int x = 0; i>0 && x < 32 ; x++)
             {
                 //int z = (1 << x); //MEM care
                 if (i & (1 << x) )
